requirements.txt
fastapi
uvicorn
python-multipart
Pillow
PyMuPDF
docx2pdf
pillow-heif
svglib
requests
PySide6
python-docx
